COM-1041 Fix after PR comments
