--- src/routes/db.js
@@ -1,16 +1,0 @@
-
-'use-strict';
-
-const { seedDb } = require('../controllers/dbController');
-
-exports.plugin = {
-  name: 'routes-db',
-  register: async (server) => {
-    server.route({
-      method: 'GET',
-      path: '/seed',
-      options: { auth: false },
-      handler: seedDb,
-    });
-  },
-};
