+++ src/routes/preHandlers/endToEnd.js
@@ -1,0 +1,7 @@
+const Boom = require('@hapi/boom');
+
+exports.authorizeDatabaseSeed = async () => {
+  if (process.env.NODE_ENV !== 'test') throw Boom.forbidden();
+
+  return null;
+};
